requirements.txt
streamlit
pandas
scikit-learn
openai
python-dotenv
sentence-transformers
numpy
groq
